Use single square brackets in shell commands
My previous commit tagged as (v2.0.1) fixed a logic error in
the shell script for determing if a branch contains merge commits,
however it uses double square brackets which aren't available in the
shell that executes the commands.

This commit changes the double square brackets for single square
brackets.

Related: ONYX-24026