Compensate for checkout@v2 checking out merge ref
This action needs the git repo to be checked out at the head of the
PR branch, not at the merge commit of the PR branch and target branch.
checkout@v2 does checkout the merge commit, so this commit switches
back to the head of the PR branch before checking for merges.